+++ engine/src/main/java/nl/inl/blacklab/search/results/Facets.java
@@ -33,7 +33,7 @@ public class Facets implements SearchResult {
     }
 
     @Override
-    public int numberOfResultObjects() {
+    public long numberOfResultObjects() {
         return resultObjects;
     }
