+++ engine/src/main/java/nl/inl/blacklab/search/results/HitsFromQuery.java
@@ -171,7 +171,7 @@ public class HitsFromQuery extends Hits {
      *            negative, reads all hits
      */
     @Override
-    protected void ensureResultsRead(int number) {
+    protected void ensureResultsRead(long number) {
         try {
             // Prevent locking when not required
             if (sourceSpansFullyRead || (number >= 0 && hitsArrays.size() > number))
